+++ R/print.ddf.gof.R
@@ -83,12 +83,17 @@ print.ddf.gof <- function(x, digits=3, ...){
                   " with ", gof$chisquare$chi2$df,
                   " degrees of freedom\n", sep=""))
       }
-
-      cat(paste("\n\nTotal chi-square = ",
-                format(gof$chisquare$pooled.chi$chisq, digits=5),
-                "  P = ", format(gof$chisquare$pooled.chi$p, digits=5),
-                " with ", gof$chisquare$pooled.chi$df, " degrees of freedom\n",
-                sep=""))
+      
+      if(!is.na(gof$chisquare$pooled.chi$p)){
+        cat(paste("\n\nTotal chi-square = ",
+                  format(gof$chisquare$pooled.chi$chisq, digits=5),
+                  "  P = ", format(gof$chisquare$pooled.chi$p, digits=5),
+                  " with ", gof$chisquare$pooled.chi$df, " degrees of freedom\n",
+                  sep=""))
+      }else{
+        cat("\nTotal chi-square: No degrees of freedom for test\n")
+      }
+      
     }
   }